Move helper functions to model utils
