+++ credoai/artifacts/model/classification_model.py
@@ -2,6 +2,7 @@
 from .base_model import Model
 
 from credoai.utils import global_logger
+from utils.model_utils import clf_handle_keras, clf_handle_torch
 
 import numpy as np
 import pandas as pd
@@ -160,160 +161,6 @@ class ClassificationModel(Model):
             raise Exception(
                 "`predict` function required for custom model {self.name}. None specified."
             )
-
-
-def clf_handle_keras(model_like):
-    predict_obj = None
-    predict_proba_obj = None
-    clf_type = "BINARY_CLASSIFICATION"
-
-    pred_func = getattr(model_like, "predict", getattr(model_like, "call"))
-
-    if model_like.layers[-1].output_shape == (None, 1):
-        # Assumes sigmoid -> probabilities need to be rounded
-        predict_obj = lambda x: pred_func(x).round()
-        # Single-output sigmoid is binary by definition
-        clf_type = "BINARY_CLASSIFICATION"
-    else:
-        # Assumes softmax -> probabilities need to be argmaxed
-        predict_obj = lambda x: np.argmax(pred_func(x), axis=1)
-        if model_like.layers[-1].output_shape[1] == 2:
-            clf_type = "BINARY_CLASSIFICATION"
-        else:
-            clf_type = "MULTICLASS_CLASSIFICATION"
-
-    if model_like.layers[-1].output_shape == (None, 2):
-        predict_proba_obj = lambda x: pred_func(x)[:, 1]
-    elif (
-        len(model_like.layers[-1].output_shape) == 2
-        and model_like.layers[-1].output_shape[1] == 1
-    ):
-        # Sigmoid -> needs to be (n_samples, ) to work with sklearn metrics
-        predict_proba_obj = lambda x: np.reshape(pred_func(x), (-1, 1))
-    elif (
-        len(model_like.layers[-1].output_shape) == 2
-        and model_like.layers[-1].output_shape[1] > 2
-    ):
-        predict_proba_obj = pred_func
-    else:
-        pass
-        # predict_proba is not valid (for now) --> this would correspond to multi-dimensional outputs or something similarly weird
-
-    return predict_obj, predict_proba_obj, clf_type
-
-
-def clf_handle_torch(model_like):
-    predict_obj = None
-    predict_proba_obj = None
-    clf_type = "BINARY_CLASSIFICATION"
-
-    pred_func = getattr(model_like, "forward", None)
-    if pred_func is None:
-        raise ValueError("Model should have a `forward` method to perform predictions.")
-
-    output_activation = getattr(model_like, "output_activation", "sigmoid")
-    if output_activation not in {"sigmoid", "softmax"}:
-        logging.warning(
-            "Invalid output activation function provided. Using sigmoid as the default activation function."
-        )
-        output_activation = "sigmoid"
-
-    with torch.no_grad():
-        dummy_input = torch.randn(1, *model_like.input_shape)
-        output = model_like(dummy_input)
-
-    output_shape = output.shape
-
-    if output_activation == "sigmoid" and output_shape[-1] == 1:
-
-        def pred(x):
-            if isinstance(x, torch.utils.data.dataloader.DataLoader):
-                preds = []
-                for data, _ in x:
-                    preds.append(torch.round(torch.sigmoid(pred_func(data))).numpy())
-                return np.concatenate(preds, axis=0)
-            else:
-                return torch.round(torch.sigmoid(pred_func(torch.tensor(x)))).numpy()
-
-        predict_obj = pred
-        clf_type = "BINARY_CLASSIFICATION"
-    elif output_activation == "softmax":
-
-        def pred(x):
-            if isinstance(x, torch.utils.data.dataloader.DataLoader):
-                preds = []
-                for data, _ in x:
-                    preds.append(
-                        torch.argmax(
-                            torch.softmax(pred_func(data), dim=-1), dim=-1
-                        ).numpy()
-                    )
-                return np.concatenate(preds, axis=0)
-            else:
-                return torch.argmax(
-                    torch.softmax(pred_func(torch.tensor(x)), dim=-1), dim=-1
-                ).numpy()
-
-        predict_obj = pred
-        clf_type = (
-            "BINARY_CLASSIFICATION"
-            if output_shape[-1] == 2
-            else "MULTICLASS_CLASSIFICATION"
-        )
-
-    if output_activation == "sigmoid" and output_shape[-1] == 1:
-
-        def pred_prob(x):
-            if isinstance(x, torch.utils.data.dataloader.DataLoader):
-                probs = []
-                for data, _ in x:
-                    probs.append(torch.sigmoid(pred_func(data)).numpy())
-                return np.concatenate(probs, axis=0)
-            else:
-                return torch.sigmoid(pred_func(torch.tensor(x))).numpy()
-
-        predict_proba_obj = pred_prob
-    elif output_activation == "softmax":
-        if output_shape[-1] == 2:
-
-            def pred_prob(x):
-                if isinstance(x, torch.utils.data.dataloader.DataLoader):
-                    probs = []
-                    for data, _ in x:
-                        probs.append(
-                            torch.softmax(pred_func(data), dim=-1)
-                            .detach()
-                            .numpy()[:, 1]
-                        )
-                    return np.concatenate(probs, axis=0)
-                else:
-                    return (
-                        torch.softmax(pred_func(torch.tensor(x)), dim=-1)
-                        .detach()
-                        .numpy()[:, 1]
-                    )
-
-            predict_proba_obj = pred_prob
-        else:
-
-            def pred_prob(x):
-                if isinstance(x, torch.utils.data.dataloader.DataLoader):
-                    probs = []
-                    for data, _ in x:
-                        probs.append(
-                            torch.softmax(pred_func(data), dim=-1).detach().numpy()
-                        )
-                    return np.concatenate(probs, axis=0)
-                else:
-                    return (
-                        torch.softmax(pred_func(torch.tensor(x)), dim=-1)
-                        .detach()
-                        .numpy()
-                    )
-
-            predict_proba_obj = pred_prob
-
-    return predict_obj, predict_proba_obj, clf_type
 
 
 class DummyClassifier:
